refactor: move SecretStoreRegistry to the secret-store-api module
The engine will resolve cached secret values through the registry on job
activation, and zeebe modules must not depend on the dist module.
